requirements.txt
torch==1.13.1+cu117
torchvision==0.14.1+cu117
transformers==4.30.0
huggingface-hub==0.36.2
rdkit==2025.9.1
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.6.1
pillow==11.3.0
tqdm==4.67.1
PyYAML==6.0.3